New Obs and Code Updates
Changed order from 2 to 3 on polyfit per an error that appeared. I'm guessing it's a change in numpy with the cov(variance) switch set.

I discovered error in spreadsheet calculation for Mu Cyg for the epoch. It was corrected to 1958 from 1962.5. I'm not sure where that error came from. It doesn't appear to be cut and paste from another binary. This has the potential to affect other stars measured during 20080915UT, 20131026UT, and of course 20180923UT.

Instead of using the weighted computed plate scale, I used a fixed plate scale of 0.175”/pix based on Jupiter and Saturn Observations fit to moon locations in WinJUPOS with the same imaging configuration.  The uncertainty is about +/-0.003 "/pix.

I added new observations for Eps Lyr, BU648, Zet Her, Mu Cyg, Del Cyg, and Lam Cyg.